contrib/gcc-12: Patch options to use DWARF v4 by default
Because our GDB in base is 7.6 and is too old to read the DWARF v5 that
is used by GCC 12.5 by default.  In addition, there is no longer a
configure option to override the default DWARF version.  Therefore, we
decided to patch the GCC source code.

We could tweak the world/kernel makefiles to add the
'-gdwarf-4 -gstrict-dwarf' options, but that couldn't solve the issue
that a user cannot use the base GDB to debug one's own binary built
outside the world/kernel.

Discussed-with: swildner